Revert "Zabbix API updates"
