+++ playbooks/adhoc/zabbix_setup/create_template.yml
@@ -1,0 +1,57 @@
+---
+- debug: var=ctp_template
+
+- name: Create Template
+  zbx_template:
+    server: "{{ ctp_zserver }}"
+    user: "{{ ctp_zuser }}"
+    password: "{{ ctp_zpassword }}"
+    name: "{{ ctp_template.name }}"
+  register: ctp_created_template
+
+- debug: var=ctp_created_template
+
+#- name: Create Application
+#  zbxapi:
+#    server: "{{ ctp_zserver }}"
+#    user: "{{ ctp_zuser }}"
+#    password: "{{ ctp_zpassword }}"
+#    zbx_class: Application
+#    state: present
+#    params:
+#      name: "{{ ctp_template.application.name}}"
+#      hostid: "{{ ctp_created_template.results[0].templateid }}"
+#      search:
+#        name: "{{ ctp_template.application.name}}"
+#  register: ctp_created_application
+
+#- debug: var=ctp_created_application
+
+- name: Create Items
+  zbx_item:
+    server: "{{ ctp_zserver }}"
+    user: "{{ ctp_zuser }}"
+    password: "{{ ctp_zpassword }}"
+    key: "{{ item.key }}"
+    name: "{{ item.name | default(item.key, true) }}"
+    value_type: "{{ item.value_type | default('int') }}"
+    template_name: "{{ ctp_template.name }}"
+  with_items: ctp_template.zitems
+  register: ctp_created_items
+
+#- debug: var=ctp_created_items
+
+- name: Create Triggers
+  zbx_trigger:
+    server: "{{ ctp_zserver }}"
+    user: "{{ ctp_zuser }}"
+    password: "{{ ctp_zpassword }}"
+    description: "{{ item.description }}"
+    expression: "{{ item.expression }}"
+    priority: "{{ item.priority }}"
+  with_items: ctp_template.ztriggers
+  when: ctp_template.ztriggers is defined
+
+#- debug: var=ctp_created_triggers
+
+
