+++ playbooks/adhoc/zabbix_setup/filter_plugins
@@ -1,0 +1,1 @@
+../../../filter_plugins
