+++ playbooks/adhoc/zabbix_setup/roles
@@ -1,0 +1,1 @@
+../../../roles/
