+++ playbooks/adhoc/zabbix_setup/vars/template_heartbeat.yml
@@ -1,11 +1,9 @@
 ---
 g_template_heartbeat:
   name: Template Heartbeat
-  zapplications:
-  - Heartbeat
   zitems:
   - name: Heartbeat Ping
-    application: Heartbeat
+    hostid:
     key: heartbeat.ping
   ztriggers:
   - description: 'Heartbeat.ping has failed on {HOST.NAME}'
