+++ playbooks/adhoc/zabbix_setup/vars/template_host.yml
