+++ playbooks/adhoc/zabbix_setup/vars/template_node.yml
