--- roles/os_zabbix/defaults/main.yml
@@ -1,1 +1,0 @@
----
