--- roles/os_zabbix/handlers/main.yml
@@ -1,1 +1,0 @@
----
