+++ roles/os_zabbix/library/get_drule.yml
@@ -1,0 +1,115 @@
+---
+# This is a test playbook to create one of each of the zabbix ansible modules.
+# ensure that the zbxapi module is installed
+# ansible-playbook test.yml
+- name: Test zabbix ansible module
+  hosts: localhost
+  gather_facts: no
+  vars:
+#zbx_server: https://localhost/zabbix/api_jsonrpc.php
+#zbx_user: Admin
+#zbx_password: zabbix
+
+  pre_tasks:
+  - name: Template Discovery rules
+    zbx_template:
+      server: "{{ zbx_server }}"
+      user: "{{ zbx_user }}"
+      password: "{{ zbx_password }}"
+      name: 'Template App HaProxy'
+      state: list
+    register: template_output
+
+  - debug: var=template_output
+
+  - name: Discovery rules
+    zbx_discovery_rule:
+      server: "{{ zbx_server }}"
+      user: "{{ zbx_user }}"
+      password: "{{ zbx_password }}"
+      name: 'haproxy.discovery sender'
+      state: list
+    register: drule
+
+  - debug: var=drule
+
+#  - name: Create an application
+#    zbx_application:
+#      server: "{{ zbx_server }}"
+#      user: "{{ zbx_user }}"
+#      password: "{{ zbx_password }}"
+#      name: 'Test App'
+#      template_name: "test template"
+#    register: item_output
+#
+#  - name: Create an item
+#    zbx_item:
+#      server: "{{ zbx_server }}"
+#      user: "{{ zbx_user }}"
+#      password: "{{ zbx_password }}"
+#      name: 'test item'
+#      key: 'kenny.item.1'
+#      applications:
+#      - 'Test App'
+#      template_name: "test template"
+#    register: item_output
+#
+#  - debug: var=item_output
+#
+#  - name: Create an trigger
+#    zbx_trigger:
+#      server: "{{ zbx_server }}"
+#      user: "{{ zbx_user }}"
+#      password: "{{ zbx_password }}"
+#      expression: '{test template:kenny.item.1.last()}>2'
+#      description: 'Kenny desc'
+#    register: trigger_output
+#
+#  - debug: var=trigger_output
+#
+#  - name: Create a hostgroup
+#    zbx_hostgroup:
+#      server: "{{ zbx_server }}"
+#      user: "{{ zbx_user }}"
+#      password: "{{ zbx_password }}"
+#      name: 'kenny hostgroup'
+#    register: hostgroup_output
+#
+#  - debug: var=hostgroup_output
+#
+#  - name: Create a host
+#    zbx_host:
+#      server: "{{ zbx_server }}"
+#      user: "{{ zbx_user }}"
+#      password: "{{ zbx_password }}"
+#      name: 'kenny host'
+#      template_names:
+#      - test template
+#      hostgroup_names:
+#      - kenny hostgroup
+#    register: host_output
+#
+#  - debug: var=host_output
+#
+#  - name: Create a usergroup
+#    zbx_usergroup:
+#      server: "{{ zbx_server }}"
+#      user: "{{ zbx_user }}"
+#      password: "{{ zbx_password }}"
+#      name: kenny usergroup
+#      rights:
+#      - 'kenny hostgroup': rw
+#    register: usergroup_output
+#
+#  - debug: var=usergroup_output
+#
+#  - name: Create a user
+#    zbx_user:
+#      server: "{{ zbx_server }}"
+#      user: "{{ zbx_user }}"
+#      password: "{{ zbx_password }}"
+#      alias: kwoodson
+#      state: list
+#    register: user_output
+#
+#  - debug: var=user_output
