+++ roles/os_zabbix/library/zbx_discoveryrule.py
