+++ roles/os_zabbix/library/zbx_host.py
