+++ roles/os_zabbix/library/zbx_item.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python
 '''
-ansible module for zabbix users
+ Ansible module for zabbix items
 '''
 # vim: expandtab:tabstop=4:shiftwidth=4
 #
-#   Zabbix user ansible module
+#   Zabbix item ansible module
 #
 #
 #   Copyright 2015 Red Hat Inc.
@@ -21,6 +21,7 @@ ansible module for zabbix users
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 #
+
 # This is in place because each module looks similar to each other.
 # These need duplicate code as their behavior is very similar
 # but different for each zabbix class.
@@ -40,55 +41,51 @@ def exists(content, key='result'):
 
     return True
 
-def get_usergroups(zapi, usergroups):
-    ''' Get usergroups
+def get_value_type(value_type):
     '''
-    ugroups = []
-    for ugr in usergroups:
-        content = zapi.get_content('usergroup',
-                                   'get',
-                                   {'search': {'name': ugr},
-                                    #'selectUsers': 'userid',
-                                    #'getRights': 'extend'
-                                   })
-        if content['result']:
-            ugroups.append({'usrgrpid': content['result'][0]['usrgrpid']})
-
-    return ugroups or None
-
-def get_usertype(user_type):
+    Possible values:
+    0 - numeric float;
+    1 - character;
+    2 - log;
+    3 - numeric unsigned;
+    4 - text
     '''
-    Determine zabbix user account type
+    vtype = 0
+    if 'int' in value_type:
+        vtype = 3
+    elif 'char' in value_type:
+        vtype = 1
+    elif 'str' in value_type:
+        vtype = 4
+
+    return vtype
+
+def get_app_ids(zapi, application_names):
+    ''' get application ids from names
     '''
-    if not user_type:
-        return None
-
-    utype = 1
-    if 'super' in user_type:
-        utype = 3
-    elif 'admin' in user_type or user_type == 'admin':
-        utype = 2
-
-    return utype
+    app_ids = []
+    for app_name in application_names:
+        content = zapi.get_content('application', 'get', {'search': {'name': app_name}})
+        if content.has_key('result'):
+            app_ids.append(content['result'][0]['applicationid'])
+    return app_ids
 
 def main():
     '''
-    ansible zabbix module for users
+    ansible zabbix module for zbx_item
     '''
-
-    ##def user(self, name, state='present', params=None):
 
     module = AnsibleModule(
         argument_spec=dict(
             server=dict(default='https://localhost/zabbix/api_jsonrpc.php', type='str'),
             user=dict(default=None, type='str'),
             password=dict(default=None, type='str'),
-            alias=dict(default=None, type='str'),
             name=dict(default=None, type='str'),
-            surname=dict(default=None, type='str'),
-            user_type=dict(default=None, type='str'),
-            passwd=dict(default=None, type='str'),
-            user_groups=dict(default=[], type='list'),
+            key=dict(default=None, type='str'),
+            template_name=dict(default=None, type='str'),
+            zabbix_type=dict(default=2, type='int'),
+            value_type=dict(default='int', type='str'),
+            applications=dict(default=[], type='list'),
             debug=dict(default=False, type='bool'),
             state=dict(default='present', type='str'),
         ),
@@ -96,43 +93,49 @@ def main():
     )
 
     user = module.params.get('user', os.environ['ZABBIX_USER'])
-    password = module.params.get('password', os.environ['ZABBIX_PASSWORD'])
+    passwd = module.params.get('password', os.environ['ZABBIX_PASSWORD'])
 
-    zapi = ZabbixAPI(ZabbixConnection(module.params['server'], user, password, module.params['debug']))
+    zapi = ZabbixAPI(ZabbixConnection(module.params['server'], user, passwd, module.params['debug']))
 
-    ## before we can create a user media and users with media types we need media
-    zbx_class_name = 'user'
-    idname = "userid"
-    alias = module.params['alias']
+    #Set the instance and the template for the rest of the calls
+    zbx_class_name = 'item'
+    idname = "itemid"
     state = module.params['state']
+    key = module.params['key']
+    template_name = module.params['template_name']
+
+    content = zapi.get_content('template', 'get', {'search': {'host': template_name}})
+    templateid = None
+    if content['result']:
+        templateid = content['result'][0]['templateid']
+    else:
+        module.exit_json(changed=False,
+                         results='Error: Could find template with name %s for item.' % template_name,
+                         state="Unkown")
 
     content = zapi.get_content(zbx_class_name,
                                'get',
-                               {'output': 'extend',
-                                'search': {'alias': alias},
-                                "selectUsrgrps": 'usergrpid',
+                               {'search': {'key_': key},
+                                'selectApplications': 'applicationid',
                                })
     if state == 'list':
         module.exit_json(changed=False, results=content['result'], state="list")
 
     if state == 'absent':
-        if not exists(content) or len(content['result']) == 0:
+        if not exists(content):
             module.exit_json(changed=False, state="absent")
 
         content = zapi.get_content(zbx_class_name, 'delete', [content['result'][0][idname]])
         module.exit_json(changed=True, results=content['result'], state="absent")
 
     if state == 'present':
-        params = {'alias': alias,
-                  'passwd': module.params['passwd'],
-                  'usrgrps': get_usergroups(zapi, module.params['user_groups']),
-                  'name': module.params['name'],
-                  'surname': module.params['surname'],
-                  'type': get_usertype(module.params['user_type']),
+        params = {'name': module.params.get('name', module.params['key']),
+                  'key_': key,
+                  'hostid': templateid,
+                  'type': module.params['zabbix_type'],
+                  'value_type': get_value_type(module.params['value_type']),
+                  'applications': get_app_ids(zapi, module.params['applications']),
                  }
-
-        # Remove any None valued params
-        _ = [params.pop(key, None) for key in params.keys() if params[key] is None]
 
         if not exists(content):
             # if we didn't find it, create it
@@ -143,15 +146,8 @@ def main():
         differences = {}
         zab_results = content['result'][0]
         for key, value in params.items():
-            if key == 'passwd':
-                differences[key] = value
 
-            elif key == 'usrgrps':
-                # this must be done as a list of ordered dictionaries fails comparison
-                if not all([True for _ in zab_results[key][0] if _ in value[0]]):
-                    differences[key] = value
-
-            elif zab_results[key] != value and zab_results[key] != str(value):
+            if zab_results[key] != value and zab_results[key] != str(value):
                 differences[key] = value
 
         if not differences:
