+++ roles/os_zabbix/library/zbx_itemprototype.py
