--- roles/os_zabbix/meta/main.yml
@@ -1,9 +1,0 @@
----
-galaxy_info:
-  author: OpenShift
-  description:  ZabbixAPI
-  company: Red Hat, Inc
-  license: ASL 2.0
-  min_ansible_version: 1.2
-dependencies:
-- lib_zabbix
