--- roles/os_zabbix/tasks/clean_zabbix.yml
@@ -1,47 +1,0 @@
----
-- name: CLEAN List template for heartbeat
-  zbx_template:
-    server: "{{ server }}"
-    user: "{{ user }}"
-    password: "{{ password }}"
-    state: list
-    name: 'Template Heartbeat'
-  register: templ_heartbeat
-
-- name: CLEAN List template app zabbix server
-  zbx_template:
-    server: "{{ server }}"
-    user: "{{ user }}"
-    password: "{{ password }}"
-    state: list
-    name: 'Template App Zabbix Server'
-  register: templ_zabbix_server
-
-- name: CLEAN List template app zabbix server
-  zbx_template:
-    server: "{{ server }}"
-    user: "{{ user }}"
-    password: "{{ password }}"
-    state: list
-    name: 'Template App Zabbix Agent'
-  register: templ_zabbix_agent
-
-- name: CLEAN List all templates
-  zbx_template:
-    server: "{{ server }}"
-    user: "{{ user }}"
-    password: "{{ password }}"
-    state: list
-  register: templates
-
-- debug: var=templ_heartbeat.results
-
-- name: Remove templates if heartbeat template is missing
-  zbx_template:
-    server: "{{ server }}"
-    user: "{{ user }}"
-    password: "{{ password }}"
-    name: "{{ item }}"
-    state: absent
-  with_items: "{{ templates.results | difference(templ_zabbix_agent.results) | difference(templ_zabbix_server.results) | oo_collect('host') }}"
-  when:  templ_heartbeat.results | length == 0
