--- roles/os_zabbix/tasks/create_template.yml
@@ -1,56 +1,0 @@
----
-- debug: var=template
-
-- name: Template Create Template
-  zbx_template:
-    server: "{{ server }}"
-    user: "{{ user }}"
-    password: "{{ password }}"
-    name: "{{ template.name }}"
-  register: created_template
-
-- debug: var=created_template
-
-- name: Create Application
-  zbx_application:
-    server: "{{ server }}"
-    user: "{{ user }}"
-    password: "{{ password }}"
-    name: "{{ item }}"
-    template_name: "{{ template.name }}"
-  with_items: template.zapplications
-  register: created_application
-  when: template.zapplications is defined
-
-- debug: var=created_application
-
-- name: Create Items
-  zbx_item:
-    server: "{{ server }}"
-    user: "{{ user }}"
-    password: "{{ password }}"
-    key: "{{ item.key }}"
-    name: "{{ item.name | default(item.key, true) }}"
-    value_type: "{{ item.value_type | default('int') }}"
-    template_name: "{{ template.name }}"
-    applications: "{{ item.application }}"
-  with_items: template.zitems
-  register: created_items
-  when: template.zitems is defined
-
-#- debug: var=ctp_created_items
-
-- name: Create Triggers
-  zbx_trigger:
-    server: "{{ server }}"
-    user: "{{ user }}"
-    password: "{{ password }}"
-    description: "{{ item.description }}"
-    expression: "{{ item.expression }}"
-    priority: "{{ item.priority }}"
-  with_items: template.ztriggers
-  when: template.ztriggers is defined
-
-#- debug: var=ctp_created_triggers
-
-
