--- roles/os_zabbix/tasks/create_user.yml
@@ -1,11 +1,0 @@
----
-- name: Update zabbix credentialss for a user
-  zbx_user:
-    server: "{{ ozb_server }}"
-    user: "{{ ozb_user }}"
-    password: "{{ ozb_password }}"
-    alias: "{{ ozb_username }}"
-    passwd: "{{ ozb_new_password | default(ozb_password, true) }}"
-  register: user
-
-- debug: var=user.results
