--- roles/os_zabbix/tasks/main.yml
@@ -1,30 +1,0 @@
----
-- name: Main List all templates
-  zbx_template:
-    server: "{{ ozb_server }}"
-    user: "{{ ozb_user }}"
-    password: "{{ ozb_password }}"
-    state: list
-  register: templates
-
-- debug: var=templates
-
-- include_vars: template_heartbeat.yml
-- include_vars: template_os_linux.yml
-
-- name: Include Template Heartbeat
-  include: create_template.yml
-  vars:
-    template: "{{ g_template_heartbeat }}"
-    server: "{{ ozb_server }}"
-    user: "{{ ozb_user }}"
-    password: "{{ ozb_password }}"
-
-#- name: Include Template os_linux
-#  include: create_template.yml
-#  vars:
-#    template: "{{ g_template_os_linux }}"
-#    server: "{{ ozb_server }}"
-#    user: "{{ ozb_user }}"
-#    password: "{{ ozb_password }}"
-
