--- roles/os_zabbix/vars/main.yml
@@ -1,1 +1,0 @@
----
